+++ LoanCalculator/Models/MortgageLoan.cs
@@ -4,7 +4,7 @@
     {
         public double rate = 3.5;
 
-        public MortgageLoan(int amount, int years) : base(amount, 3.5, years)
+        public MortgageLoan(decimal amount, int years) : base(amount, 3.5, years)
         {
 
         }
